renamed statistics.rs to metrics.rs #706
